+++ realm/realm-library/src/main/java/io/realm/BaseRealm.java
@@ -817,12 +817,16 @@ abstract class BaseRealm implements Closeable {
      *
      *   private Realm realm = null;
      *   private RealmAsyncTask realmAsyncTask;
+     *   private static RealmConfiguration config = new RealmConfiguration.Builder()
+     *     .schema(42)
+     *     .migration(new MyMigration()) // Potentially lengthy migration
+     *     .build();
      *
      *   \@Override
      *   protected void onCreate(Bundle savedInstanceState) {
      *     super.onCreate(savedInstanceState);
      *     setContentView(R.layout.layout_main);
-     *     realmAsyncTask = Realm.getDefaultInstanceAsync(new Callback() {
+     *     realmAsyncTask = Realm.getInstanceAsync(config, new Callback() {
      *         \@Override
      *         public void onSuccess(Realm realm) {
      *             if (isDestroyed()) {